LayoutTests/media/video-dom-loopend-expected.txt

Test that video loops when time reaches the value specified by the loopEnd DOM attribute.

TEST(video.loopEnd==0.5) OK
TEST(video.getAttribute('loopend') == '0.5s') OK
RUN(video.play())
EVENT(timeupdate)
EVENT(timeupdate)
TEST(video.currentLoop == 1) OK
TEST(video.currentTime.toFixed(1) == 0) OK
END OF TEST
